Refactor code and add new OrderItemStockLocationAssignment functionality
The main LinnMacroCustomer program has been updated to remove some redundant code and introduce a new method for executing the OrderItemStockLocationAssignment. Some of the changes include the creation of a new LinnworksMacro class and corresponding project within the OrderItemStockLocationAssignment directory, as well as modifications to how macros are set up and executed within the LinnMacroCustomer program. These changes aim to enhance the functionality and improve the organization of the code.